+++ NEWS.md
@@ -4,6 +4,7 @@
 # moderndive 0.1.0
 
 * Added three `get_regression_*` functions
-
-
+* Added datasets:
+    + `pennies`: 800 pennies to be treated as a population from which to sample
+    + `ball_samples` 10 groups sampled n=50 balls from this [tub](https://raw.githubusercontent.com/ismayc/moderndiver-book/master/images/sampling1.jpg) and counted the number red