+++ NEWS.md
@@ -1,3 +1,5 @@
+# moderndive 0.2.0
+* Changed the functions to not include the `lm` call based on feedback from CRAN
 
 # moderndive 0.1.0